\name{loadInputFile}
\alias{loadInputFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Load the input file}
\description{
  Internal function.
}
\usage{
loadInputFile(InputFile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{InputFile}{file name}
}
\keyword{file}

