\name{pc_graphs_GEMp}
\alias{pc_graphs_GEMp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot ancestral structures}
\description{
  The function makes .pdf and .ps plots based on the significant 
eigenvectors from the SpectralGEM function.
}
\usage{
pc_graphs_GEMp(ext)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ext}{the file extension generated by the identifier and the 
stage. The values are printed out in the R console after runing 
SpectralGEM().}
}
\details{
  The function plots pairs of principle componets corresponding to the 
significant eigenvectors.
}
\value{
  The plots are in .pdf form and .ps form under the current directory.
}
\references{ 
http://wpicr.wpic.pitt.edu/WPICCompGen/Spectral-GEM/GEM+.htm}
\author{Ann Lee, Diana Luca, Bert Klei, Bernie Devlin, and 
Kathryn  Roeder
  
       Maintainer: Jing Wu jwu@stat.cmu.edu}
\examples{
# pc_graphs_GEMp("smal1")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{dplot}

