\name{plotLam}
\alias{plotLam}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot the eigenvalues}
\description{
The function plots the eigenvalues the main fortran program produces. It
helps user to determine the number of dimensions (mdim) to be used in the program.
}
\usage{
plotLam(ext)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ext}{The identifier}
}
\details{
 The function generates a plot of all the eigenvalues.
}
\value{
}
\references{http://wpicr.wpic.pitt.edu/WPICCompGen/Spectral-GEM/GEM+.htm }
\author{Ann Lee, Diana Luca, Bert Klei, Bernie Devlin, and Kathryn
  Roeder
  
        Maintainer: Jing Wu jwu@stat.cmu.edu}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ file }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
